New features added
Added queue functionality to the bot that allows users to add many songs to a queue and the bot will automatically cycle through them. A clear queue functionality is added to clear the queue, a skip functionality is also added and a basic interface displaying currently playing song and queued songs is also added.